refactor(naming): harmonize file type discriminants
Establish one rule for file naming: a dotted suffix (`name.<type>.ts`) marks a
genuine architectural role; a kebab name does not. Apply it to the config layer,
which mixed both conventions and used non-type suffixes.

- The ConfigType handlers are a distinct abstraction from interaction handlers
  (InteractionHandler / `.button.ts`), so the precise type is `config-handler`,
  discriminated by ConfigType: `string.config-handler.ts`, `number.…`, etc.
  (was `string-config-handler.ts`). Abstract base → `config-handler.ts`,
  registry → `config-handler-registry.ts`.
- `config-edit.helpers.ts` → `config-edit.ts`: "helpers" is not a type.
- `core.messages.ts` → `core-messages.ts`: "messages" is a topic, not a type,
  so it takes the kebab form like its utils/ neighbours.

Pure renames + import-path updates; no behavior change.
