+++ src/core/config/channel.config-handler.ts
@@ -1,18 +1,19 @@
 import {
   ActionRowBuilder,
-  UserSelectMenuBuilder,
+  ChannelSelectMenuBuilder,
+  ChannelType,
   type AnySelectMenuInteraction,
   type MessageActionRowComponentBuilder,
 } from "discord.js";
 import { ConfigType } from "../../lib/config.js";
 import type { CompatibleInteraction } from "../../lib/interaction.js";
-import { EntitySelectConfigHandler } from "./entity-select-config-handler.js";
+import { EntitySelectConfigHandler } from "./entity-select.config-handler.js";
 
-export default class UserConfigHandler extends EntitySelectConfigHandler<ConfigType.USER> {
-  protected override readonly selectCustomId = "set-user-config";
+export default class ChannelConfigHandler extends EntitySelectConfigHandler<ConfigType.CHANNEL> {
+  protected override readonly selectCustomId = "set-channel-config";
 
   constructor() {
-    super(ConfigType.USER);
+    super(ConfigType.CHANNEL);
   }
 
   protected override buildSelectRow(
@@ -21,18 +22,21 @@ export default class UserConfigHandler extends EntitySelectConfigHandler<ConfigT
     minValues: number,
     maxValues: number
   ): ActionRowBuilder<MessageActionRowComponentBuilder> {
-    const menu = new UserSelectMenuBuilder()
+    const menu = new ChannelSelectMenuBuilder()
       .setCustomId(customId)
       .setPlaceholder(
-        maxValues > 1
-          ? "Sélectionnez des utilisateurs"
-          : "Sélectionnez un utilisateur"
+        maxValues > 1 ? "Sélectionnez des salons" : "Sélectionnez un salon"
+      )
+      .setChannelTypes(
+        ChannelType.GuildText,
+        ChannelType.GuildAnnouncement,
+        ChannelType.GuildVoice
       )
       .setMinValues(minValues)
       .setMaxValues(maxValues);
 
     if (currentIds.length > 0) {
-      menu.setDefaultUsers(...currentIds);
+      menu.setDefaultChannels(...currentIds);
     }
 
     return new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
@@ -43,6 +47,6 @@ export default class UserConfigHandler extends EntitySelectConfigHandler<ConfigT
   protected override isMatchingSelect(
     interaction: CompatibleInteraction
   ): interaction is AnySelectMenuInteraction {
-    return interaction.isUserSelectMenu();
+    return interaction.isChannelSelectMenu();
   }
 }
