+++ src/core/config/config-handler-registry.ts
@@ -1,11 +1,11 @@
 import { ConfigType } from "../../lib/config.js";
-import CategoryConfigHandler from "./category-config-handler.js";
-import ChannelConfigHandler from "./channel-config-handler.js";
-import type { ConfigTypeHandler } from "./config-type-handler.js";
-import NumberConfigHandler from "./number-config-handler.js";
-import RoleConfigHandler from "./role-config-handler.js";
-import StringConfigHandler from "./string-config-handler.js";
-import UserConfigHandler from "./user-config-handler.js";
+import CategoryConfigHandler from "./category.config-handler.js";
+import ChannelConfigHandler from "./channel.config-handler.js";
+import type { ConfigTypeHandler } from "./config-handler.js";
+import NumberConfigHandler from "./number.config-handler.js";
+import RoleConfigHandler from "./role.config-handler.js";
+import StringConfigHandler from "./string.config-handler.js";
+import UserConfigHandler from "./user.config-handler.js";
 
 const handlers: Record<ConfigType, ConfigTypeHandler<ConfigType> | null> = {
   [ConfigType.STRING]: new StringConfigHandler(),
