+++ src/core/config/user.config-handler.ts
@@ -1,19 +1,18 @@
 import {
   ActionRowBuilder,
-  ChannelSelectMenuBuilder,
-  ChannelType,
+  UserSelectMenuBuilder,
   type AnySelectMenuInteraction,
   type MessageActionRowComponentBuilder,
 } from "discord.js";
 import { ConfigType } from "../../lib/config.js";
 import type { CompatibleInteraction } from "../../lib/interaction.js";
-import { EntitySelectConfigHandler } from "./entity-select-config-handler.js";
+import { EntitySelectConfigHandler } from "./entity-select.config-handler.js";
 
-export default class ChannelConfigHandler extends EntitySelectConfigHandler<ConfigType.CHANNEL> {
-  protected override readonly selectCustomId = "set-channel-config";
+export default class UserConfigHandler extends EntitySelectConfigHandler<ConfigType.USER> {
+  protected override readonly selectCustomId = "set-user-config";
 
   constructor() {
-    super(ConfigType.CHANNEL);
+    super(ConfigType.USER);
   }
 
   protected override buildSelectRow(
@@ -22,21 +21,18 @@ export default class ChannelConfigHandler extends EntitySelectConfigHandler<Conf
     minValues: number,
     maxValues: number
   ): ActionRowBuilder<MessageActionRowComponentBuilder> {
-    const menu = new ChannelSelectMenuBuilder()
+    const menu = new UserSelectMenuBuilder()
       .setCustomId(customId)
       .setPlaceholder(
-        maxValues > 1 ? "Sélectionnez des salons" : "Sélectionnez un salon"
-      )
-      .setChannelTypes(
-        ChannelType.GuildText,
-        ChannelType.GuildAnnouncement,
-        ChannelType.GuildVoice
+        maxValues > 1
+          ? "Sélectionnez des utilisateurs"
+          : "Sélectionnez un utilisateur"
       )
       .setMinValues(minValues)
       .setMaxValues(maxValues);
 
     if (currentIds.length > 0) {
-      menu.setDefaultChannels(...currentIds);
+      menu.setDefaultUsers(...currentIds);
     }
 
     return new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
@@ -47,6 +43,6 @@ export default class ChannelConfigHandler extends EntitySelectConfigHandler<Conf
   protected override isMatchingSelect(
     interaction: CompatibleInteraction
   ): interaction is AnySelectMenuInteraction {
-    return interaction.isChannelSelectMenu();
+    return interaction.isUserSelectMenu();
   }
 }
